Redirect to payment from root
